Initial commit of configurable metrics endpoint for marathon discovery
This commit introduces some optional configuration to the marathon
discovery plugin to allow for configuring which port is discovered by
prometheus for scraping. This is useful if you're exporting numerous
ports from a container, and wish to not expose your metrics on a public
interface for instance, and/or if you have containers which don't have
metrics available and would like to have only those which specify this
setting indicating they have a metrics endpoint to be scraped.